requirements.txt
pandas>=2.1
numpy>=1.24
matplotlib>=3.7
seaborn>=0.13
scikit-learn>=1.3
opencv-python>=4.9
Pillow>=10.0
kaggle>=1.6
tqdm>=4.66
nltk>=3.9
wordcloud>=1.9
scikit-image>=0.22
# Deep learning
torch>=2.3
torchvision>=0.18